libs/fail: Move into internal
