+++ internal/consensus/state.go
@@ -17,8 +17,8 @@ import (
 	"github.com/cometbft/cometbft/crypto"
 	cstypes "github.com/cometbft/cometbft/internal/consensus/types"
 	cmtevents "github.com/cometbft/cometbft/internal/events"
+	"github.com/cometbft/cometbft/internal/fail"
 	sm "github.com/cometbft/cometbft/internal/state"
-	"github.com/cometbft/cometbft/libs/fail"
 	cmtjson "github.com/cometbft/cometbft/libs/json"
 	"github.com/cometbft/cometbft/libs/log"
 	cmtmath "github.com/cometbft/cometbft/libs/math"
